+++ src/CompilerCxx/Category.hs
@@ -1,5 +1,5 @@
 {- -----------------------------------------------------------------------------
-Copyright 2019-2020 Kevin P. Barry
+Copyright 2019-2021 Kevin P. Barry
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -57,7 +57,6 @@ import CompilerCxx.Naming
 import CompilerCxx.Procedure
 import Types.Builtin
 import Types.DefinedCategory
-import Types.Pragma
 import Types.Procedure
 import Types.TypeCategory
 import Types.TypeInstance